refactor(moved knexfile.js and DB directory into server directory): moved knexfile.js and DB directo
moved knexfile.js and DB directory into server directory for cleaner file structure understanding to
outside users

"re #123"
