+++ server/routes/bios.js
@@ -1,5 +1,5 @@
 const router = require("express").Router();
-const db = require('../../db/index.js')
+const db = require("../db/index.js");
 
 router.get("/:id", (req, res) => {
   const candId = req.params.id;
