+++ server/routes/candidates.js
@@ -1,5 +1,5 @@
 const router = require("express").Router();
-const db = require("../../db/index.js");
+const db = require("../db/index.js");
 
 // Test GET
 router.get("/test", (req, res) => {
